requirements.txt
# AVTO LAIF Backend - Dependencies
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Redis for caching and sessions
redis==5.2.1

# Authentication
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation and serialization
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
phonenumbers==8.13.52

# WebSocket
websockets==14.1

# Async HTTP client
httpx==0.28.1
aiohttp==3.11.11

# File handling
python-magic==0.4.27
pillow==11.1.0
aiofiles==24.1.0

# Geolocation
geopy==2.4.1

# Background tasks
celery==5.4.0
arq==0.26.1

# Utilities
python-dotenv==1.0.1
orjson==3.10.13
pendulum==3.0.0

# Logging and monitoring
structlog==24.4.0
sentry-sdk[fastapi]==2.19.2

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
factory-boy==3.3.1

# Code quality
black==24.10.0
ruff==0.8.6
mypy==1.14.1

